Your-Mood-React-App
# your-mood
Your Mood React App

This project aims to improve and put into practice development techniques in React.
How are you feeling today? will present a simple way to update the useState with buttons and images.

You can collaborate with this project to make this more attractive. The purpose is knowledge sharing.